refactor: DRY emails builder
