+++ src/builders/subscribe.ts
@@ -1,0 +1,5 @@
+import buildMail from './';
+
+export default async function buildSubscribe(params: any): Promise<unknown> {
+  return buildMail('subscribe', params);
+}
