+++ src/builders/summary.ts
@@ -1,0 +1,8 @@
+import { getProposals } from '../helpers/snapshot';
+import buildMail from './';
+
+export default async function buildSummary(params: any): Promise<unknown> {
+  params.proposals = await getProposals(params.address);
+
+  return buildMail('summary', params);
+}
