+++ src/rpc.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import { subscribe, rpcError, rpcSuccess } from './helpers/utils';
 import { send } from './helpers/mail';
+import buildSubscribe from './builders/subscribe';
 
 const router = express.Router();
 
@@ -11,11 +12,7 @@ router.post('/', async (req, res) => {
     if (method === 'snapshot.subscribe') {
       await subscribe(params.email, params.address);
 
-      await send('subscribe', {
-        to: params.email,
-        address: params.address,
-        name: 'Fabien'
-      });
+      await send(await buildSubscribe({ to: params.email, address: params.address }));
 
       return rpcSuccess(res, 'OK', id);
     }
