+++ src/sender.ts
@@ -1,19 +1,11 @@
-import { getProposals } from './helpers/snapshot';
 import { send } from './helpers/mail';
-import styles from './helpers/styles.json';
+import buildSummary from './builders/summary';
 
 async function start() {
   const to = 'fabien@bonustrack.co';
   const address = '0xeF8305E140ac520225DAf050e2f71d5fBcC543e7';
 
-  const proposals = await getProposals(address);
-
-  const result = await send('summary', {
-    to,
-    address,
-    proposals,
-    styles
-  });
+  const result = await send(await buildSummary({ to, address }));
 
   console.log('OK', result);
 }
